strip for textedit  run sql in python better

+++ py_helpers/run_sql.py
@@ -9,16 +9,26 @@ Created on Sun Jul 13 08:28:33 2025
 
 # ---- tof
 
+# ---- tof
+# --------------------
+if __name__ == "__main__":
+    import main
+    #main.main()
+# ----------------
+
+
+
 # ---- imports
 from PyQt5.QtSql import QSqlDatabase, QSqlQuery, QSqlTableModel
 # ---- end imports
 
 from stuffdb import STUFFDB_CONNECTION_NAME
+import parameters
 
 #-------------------------------
 
 # ----------------------------------------
-class KeyGen(   ):
+class KeyGenxxx(   ):
     """
     runs sql when passed a dict -- no under construction
 
@@ -100,7 +110,10 @@ class RunSql(   ):
     runs sql when passed a dict
 
     outpur to ?? a file
+    print goes back to gui
+    if run from >>py
     """
+
     def __init__(self, sql_dict ):
         """
         the usual
@@ -112,16 +125,75 @@ class RunSql(   ):
         sql_dict["db_file_name"]    =  ":memory:"
         sql_dict["db_file_name"]    = "/tmp/ramdisk/qt_sql.db"
         sql_dict["sql"]             = "select"
-        sql_dict["db_type"]             = "QSQLITE"
+        sql_dict["db_type"]         = "QSQLITE"
+
+        might want a column list for select or can we parse?
 
         """
+        self.sql_dict       = sql_dict
+        self.connect_name   = None
+        try:
+            self.go()
 
-        self.sql_dict   = sql_dict
 
-        if sql_dict["sql_type"] == "select":
+        except Exception as error:
+
+            error_message = str(error)
+            msg  = (f"Caught an erro in __init__: {error_message}")
+            print( msg )
+
+            # msg_box_msg    = "this is a message"
+            # msg_box             = QMessageBox()
+            # msg_box.setIcon( QMessageBox.Information )
+            # msg_box.setText(  msg_box_msg  )
+            # msg_box.setWindowTitle( "Sorry that is a No Go " )
+            # msg_box.setStandardButtons( QMessageBox.Ok )
+
+        finally:
+            self.finalize_db()
+            self.close_file_out()
+
+
+    def go(self,   ):
+        """
+
+
+        """
+        self.is_existing = False
+
+        key             = "output_type"
+        get_item        = self.sql_dict.get( key, None )
+        if not get_item:
+            self.sql_dict[ key ] = None
+        else:
+            self.sql_dict[ key ] = get_item.lower()
+
+        key             = "output_file_name"
+        get_item        = self.sql_dict.get( key, None )
+        if not get_item:
+            self.sql_dict[ key ] = None
+        else:
+            self.sql_dict[ key ] = get_item
+
+
+        # output_type  = sql_dict["output_type"].lower( )
+        # output_file  = sql_dict["output_file_name"]
+
+        self.columns    = None
+        self.file_out   = None   # changes if file is opened
+
+
+        sql_type   = self.sql_dict["sql_type"].lower( )
+        if  sql_type == "select":
             pass
             self.select_data()
             #self.run_sql_select   = RunSqlSelect( sql_dict )
+
+        elif sql_type ==  "delete":
+            pass
+            self.delete_data()
+
+        self.close_file_out()
 
         print( "__init__ all done ")
 
@@ -151,21 +223,23 @@ class RunSql(   ):
 
 
 
-connection_name = "my_connection"
-if QSqlDatabase.contains(connection_name):
-    self.db = QSqlDatabase.database(connection_name)
-else:
-    self.db = QSqlDatabase.addDatabase(db_type, connection_name)
-    self.db.setDatabaseName(db_file_name)
+        # connection_name = "my_connection"
+        # if QSqlDatabase.contains(connection_name):
+        #     self.db = QSqlDatabase.database(connection_name)
+        # else:
+        #     self.db = QSqlDatabase.addDatabase(db_type, connection_name)
+        #     self.db.setDatabaseName(db_file_name)
 
 
-
-        self.db         = QSqlDatabase.addDatabase( db_type  )
+        # for now use name run_sql later perhaps see if there is a dup?
+        self.db         = QSqlDatabase.addDatabase( db_type, "run_sql"  )
         self.db.setDatabaseName( db_file_name )
 
         if not self.db.open():
             msg    = f"Database Error: {self.db.lastError().databaseText()} {db_file_name =} "
             print( msg, )
+            # ore except
+            return
             # QMessageBox.critical(
             #     None,
             #     "databasenot open - Error!", msg
@@ -184,8 +258,12 @@ else:
         """
     # --------------------------------
     def finalize_db( self, ):
-        """ """
-        if not self.is_existing:
+        """
+        use if doing something useful
+        see qsql_db_access.py
+        seems to get rid of db inless pre-existing
+        """
+        if not self.is_existing and self.connect_name :
             QSqlDatabase.removeDatabase( self.connect_name )
 
 
@@ -197,6 +275,9 @@ else:
         """
         sql         = self.sql_dict["sql"]
         self.init_db()
+        self.open_file_output()
+
+        self.parse_column_list()
 
         query       = QSqlQuery( self.db  )
 
@@ -230,9 +311,10 @@ else:
 
             msg      = (f"row: {query.value(0) = }  { query.value(1) = }  {query.value(1)} ")
             print( msg )
+            self.output_to_file( msg )
 
 
-        sel.finalize_db()
+        self.finalize_db()
         print( "-------------- all done ---------------------")
 
     #-----------------------------------------------
@@ -264,58 +346,150 @@ else:
             query.addBindValue( name )
             query.addBindValue( frequency )
 
+    #-----------------------------------------------
+    def parse_column_list( self ):
+        """
+        for the select
+        require FROM in upper?
+        """
+        sql      = self.sql_dict["sql"].strip()
+
+        sql      = sql.replace( "from ",   "FROM " )
+        sql      = sql.replace( "select ", "SELECT " )
+
+        sql_split   = sql.split( "FROM" )
+        sql_select  = sql_split[0]
+        #                         1234567
+        if sql_select.startswith( "SELECT "):
+            pass
+        else:
+            return
+        sql_columns     = sql_select[ 7: ]
+        sql_columns     = sql_columns.replace( ",", " ")
+        columns         = sql_columns.split( " " )
+        # remove empty
+        columns         = [i_column for i_column in columns if i_column != ""]
+
+
+        self.columns    = columns
+
+
         # !! do we ever execute
 
     #-----------------------------------------------
     def delete_data( self ):
         """
+        no output to file but return no of deleted
 
         """
         # print_func_header( "delete_data()  not implemented" )
         # self.append_msg( tab_base.DONE_MSG )
+        sql         = self.sql_dict["sql"]
+        self.init_db()
+        # self.open_file_output()
+
+        # self.parse_column_list()
+
+        query       = QSqlQuery( self.db  )
+
+        query_ok    = True
+
+        result      = query.exec_( sql )
+
+        if not result:
+            query_ok        = False
+            error_txt       = query.lastError().text()
+            loc             = "query_exec_error_check"
+            debug_msg       = f"{loc} >>> error sql = { sql } \n lastError = {error_txt = }"
+            #logging.debug( debug_msg )
+            print( debug_msg )
+            # dialog          =  DisplaySQLError( parent = None, title = "SQL Error", msg = debug_msg )
+            # if dialog.exec_() == QDialog.Accepted:
+            #     pass
+
+            raise Exception( "sqlerror", debug_msg )
+
+        else:
+            pass
+            #rint("Query executed successfully.")
+            rows_deleted = query.numRowsAffected()
+            msg          = f"Rows deleted: {rows_deleted}"
+            print( msg )
+
+
+        # while query.next():
+        #     # a_id        = query.value(0)
+        #     # name        = query.value(1)
+        #     # frequency   = query.value(1)
+
+        #     msg      = (f"row: {query.value(0) = }  { query.value(1) = }  {query.value(1)} ")
+        #     print( msg )
+        #     self.output_to_file( msg )
+
+
+        self.finalize_db()
+        print( "-------------- all done ---------------------")
 
 
 
-class RunSqlSelectxxxx():
-    def __init__(self, run_sql ):
-        """
-        the usual
-        """
-        self.run_sql   = run_sql
-        #self.sql_dict   = sql_dict
+
+
+
+
+
+
+
+
+
+
+
+
+
 
 
 
 
     #-----------------------------------------------
-    def select_and_output( self ):
+    def open_file_output( self ):
         """
-        select data then loop through with a print
-        test qsql_utils
+        args in self.sql_dict
+        """
+        # print_func_header( "delete_data()  not implemented" )
+        # self.append_msg( tab_base.DONE_MSG )
+        output_type         = self.sql_dict["output_type"].lower( )
+        output_file_name    = self.sql_dict["output_file_name"]
+
+
+        # lets put it under output
+        output_file_name    = parameters.PARAMETERS.output_dir + "/" + output_file_name
+        output_file_name    = output_file_name.replace( "//", "/" )
+        output_file_name    = output_file_name.replace( "//", "/" )  # super sure
+
+        self.file_out       = open( output_file_name,
+                                    'w',
+                                    encoding  = "utf8",
+                                    errors    = 'ignore' )
+
+    # --------------------------------
+    def output_to_file( self, msg ):
+        """
+        self.parent_window.output_to_file( msg )
+
+
+        """
+        if self.file_out:
+            self.file_out.write( msg + "\n" )
+        #print( msg )
+
+    # --------------------------------
+    def close_file_out( self,  ):
         """
 
-        sql         =  self.sql_dict["sql"]
+        """
+        if self.file_out:
+            self.file_out.close()
+        self.file_out = None
 
-        query      = QSqlQuery(  self.run_sql.db )
-
-        query_ok   =  query_exec_error_check( query = query, sql = sql, raise_except = True )
-
-        msg      = ("query = {sql }")
-        output( msg )
-
-        while query.next():
-            # a_id        = query.value(0)
-            # name        = query.value(1)
-            # frequency   = query.value(2)
-
-            msg      = (f"ID: {query.value(0)}  {query.value(0)}  {query.value(0) }  ")
-            output( msg )
-
-
-
-def output( a_string ):
-    """ """
-    print( a_string )
